src/digit-id-bot/requirements.txt
pyTelegramBotAPI==4.28.0
python-dotenv==1.1.1
pytz==2025.2
